MSVC compatibility fixes - mostly casts to avoid warnings about size truncations - fully compiles on latest MS VC++

+++ jim-aio.c
@@ -64,6 +64,7 @@
 
 #include "jim-eventloop.h"
 #include "jim-subcmd.h"
+#include "jim-win32compat.h"
 
 #define AIO_CMD_LEN 32      /* e.g. aio.handleXXXXXX */
 #define AIO_BUF_LEN 256     /* Can keep this small and rely on stdio buffering */
@@ -386,9 +387,9 @@ static int aio_cmd_read(Jim_Interp *interp, int argc, Jim_Obj *const *argv)
             readlen = AIO_BUF_LEN;
         }
         else {
-            readlen = (neededLen > AIO_BUF_LEN ? AIO_BUF_LEN : neededLen);
+            readlen = (neededLen > AIO_BUF_LEN ? AIO_BUF_LEN : (int)neededLen);
         }
-        retval = fread(buf, 1, readlen, af->fp);
+        retval = (int)fread(buf, 1, readlen, af->fp);
         if (retval > 0) {
             Jim_AppendString(interp, objPtr, buf, retval);
             if (neededLen != -1) {
@@ -478,7 +479,7 @@ static int aio_cmd_gets(Jim_Interp *interp, int argc, Jim_Obj *const *argv)
             Jim_AppendString(interp, objPtr, buf, AIO_BUF_LEN - 1);
         }
         else {
-            len = strlen(buf);
+            len = (int)strlen(buf);
 
             if (len && (buf[len - 1] == '\n')) {
                 /* strip "\n" */
@@ -723,7 +724,7 @@ static int aio_cmd_seek(Jim_Interp *interp, int argc, Jim_Obj *const *argv)
     if (Jim_GetWide(interp, argv[0], &offset) != JIM_OK) {
         return JIM_ERR;
     }
-    if (fseeko(af->fp, offset, orig) == -1) {
+    if (fseeko(af->fp, (off_t)offset, orig) == -1) {
         JimAioSetError(interp, af->filename);
         return JIM_ERR;
     }
